Recente toegevoegede stripboeken worden weergegeven op landing page + class opbouw verbeterd.
