+++ Models/ComicRepository.cs
@@ -1,0 +1,25 @@
+using System.Collections.Generic;
+using System.Linq;
+using Dapper;
+using MySql.Data.MySqlClient;
+
+namespace ProjectComicBook.Models
+{
+    public class ComicRepository
+    {
+        private MySqlConnection Connect()
+        {
+            return new MySqlConnection("SERVER=localhost;DATABASE=projectcomicbook;user=root;PASSWORD=;SSL Mode=None;");
+        }
+    
+        public List<Comic> Search(string search)
+        {
+            using var connection = Connect();
+            var comic = connection.Query<Comic>("SELECT * FROM comicbook WHERE titel LIKE @search OR description LIKE @search", new
+            {
+                search = "%" + search + "%"
+            });
+            return comic.ToList();
+        }
+    }
+}
