--- Pages/Comicsbook.cs
@@ -1,14 +1,0 @@
-namespace ProjectComicBook.Pages;
-
-public class Comic
-{
-    public int comicbookID { get; set; }
-    public int serie_ID { get; set; }
-    public string titel { get; set; }
-    public string authorid { get; set; }
-    public string isbn { get; set; }
-    public string cover { get; set; }
-   
-    public string description { get; set; }
-    
-}
