+++ Pages/SearchResult.cshtml.cs
@@ -1,19 +1,22 @@
+using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
+using ProjectComicBook.Models;
 
-namespace ProjectComicBook.Pages;
-
-public class SearchResult : PageModel
+namespace ProjectComicBook.Pages
 {
-    [BindProperty (SupportsGet = true)]
-    public string Search {get;set;}
-    
-    public List<Comic> comics { get; set; }
-
-    public IActionResult OnGet(string Search)
+    public class SearchResult : PageModel
     {
-        var comic = new ComicRepository();
-        comics = comic.Search(Search); 
-        return Page();
+        [BindProperty (SupportsGet = true)]
+        public string Search {get;set;}
+    
+        public List<Comic> comics { get; set; }
+
+        public IActionResult OnGet(string Search)
+        {
+            var comic = new ComicRepository();
+            comics = comic.Search(Search); 
+            return Page();
+        }
     }
 }
