+++ Pages/UpdateComicbooks.cshtml.cs
@@ -1,44 +1,46 @@
+using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
-using Microsoft.Extensions.Logging;
 using ProjectComicBook.Models;
-namespace ProjectComicBook.Pages;
 
-public class UpdateComicbooks : PageModel
+namespace ProjectComicBook.Pages
 {
-    private DatabaseHandler DatabaseHandler = new DatabaseHandler();
+    public class UpdateComicbooks : PageModel
+    {
+        private DatabaseHandler DatabaseHandler = new DatabaseHandler();
     
-    public IEnumerable<dynamic> Comicbooks{ get; set; }
+        public IEnumerable<dynamic> Comicbooks{ get; set; }
 
-    public void OnGet()
-    {
-        Comicbooks = new DatabaseHandler().GetAllComicBooks();
-    }
+        public void OnGet()
+        {
+            Comicbooks = new DatabaseHandler().GetAllComicBooks();
+        }
     
-    public RedirectToPageResult OnPostDeleteComicbook()
-    {
-        string comicbookIDString = Request.Form["del-button"];
-        int comicbookID = Convert.ToInt32(comicbookIDString);
-        DatabaseHandler.DeleteComicbook(comicbookID);
-        return RedirectToPage("./updateComicbooks");
-    }
+        public RedirectToPageResult OnPostDeleteComicbook()
+        {
+            string comicbookIDString = Request.Form["del-button"];
+            int comicbookID = Convert.ToInt32(comicbookIDString);
+            DatabaseHandler.DeleteComicbook(comicbookID);
+            return RedirectToPage("./updateComicbooks");
+        }
     
-    public RedirectToPageResult OnPostUpdateComicbook()
-    {
-        string comicbookIDString = Request.Form["upd-button"];
-        int comicbookID = Convert.ToInt32(comicbookIDString);
-        var title = Request.Form["title"];
-        string descriptionComic = Request.Form["descriptionComic"];
-        string isbn = Request.Form["isbn"];
-        string releaseDate = Request.Form["releaseDate"];
-        string type = Request.Form["type"];
-        string pages = Request.Form["pages"];
-        int serie_ID = 1;
-        int authorID = 1;
-        int illustratorID = 1;
-        int explicitComic = 1;
-        DatabaseHandler.UpdateComicbook(comicbookID, serie_ID, authorID, illustratorID, title, descriptionComic, isbn, releaseDate, type, pages, explicitComic);
-        return RedirectToPage("./updateComicbooks");
+        public RedirectToPageResult OnPostUpdateComicbook()
+        {
+            string comicbookIDString = Request.Form["upd-button"];
+            int comicbookID = Convert.ToInt32(comicbookIDString);
+            var title = Request.Form["title"];
+            string descriptionComic = Request.Form["descriptionComic"];
+            string isbn = Request.Form["isbn"];
+            string releaseDate = Request.Form["releaseDate"];
+            string type = Request.Form["type"];
+            string pages = Request.Form["pages"];
+            int serie_ID = 1;
+            int authorID = 1;
+            int illustratorID = 1;
+            int explicitComic = 1;
+            DatabaseHandler.UpdateComicbook(comicbookID, serie_ID, authorID, illustratorID, title, descriptionComic, isbn, releaseDate, type, pages, explicitComic);
+            return RedirectToPage("./updateComicbooks");
+        }
     }
 }
