+++ ProjectComicBook.csproj.DotSettings.user
@@ -1,0 +1,3 @@
+﻿<wpf:ResourceDictionary xml:space="preserve" xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:s="clr-namespace:System;assembly=mscorlib" xmlns:ss="urn:shemas-jetbrains-com:settings-storage-xaml" xmlns:wpf="http://schemas.microsoft.com/winfx/2006/xaml/presentation">
+	<s:String x:Key="/Default/CodeInspection/WebPathMapping/IgnoredPaths/=WWWROOT_005CPROJECTCOMICBOOK_002ESTYLES_002ECSS/@EntryIndexedValue">wwwroot\ProjectComicBook.styles.css</s:String>
+	<s:String x:Key="/Default/CodeInspection/WebPathMapping/PathsInCorrectCasing/=WWWROOT_005CPROJECTCOMICBOOK_002ESTYLES_002ECSS/@EntryIndexedValue">wwwroot\ProjectComicBook.styles.css</s:String></wpf:ResourceDictionary>
